backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==0.8.11
huggingface-hub==0.16.4
torch==2.0.1
transformers==4.30.2
gunicorn==21.2.0
